Move line comment options from comments/ to options/
Relocated line-comment-options.js from src/plugin/src/comments to src/plugin/src/options so the option normalization utilities live alongside the other plugin option helpers.
