+++ src/plugin/src/options/line-comment-options.js
